# Improve asymptotic performance of binary treemap