@mbostock 2016-05-27
Exciting! This looks promising. Will try to review soon. Thank you for the contribution.