@mbostock 2016-06-07
I polished it a tiny bit in f5ba7f17db7c98aa1aaf7318cb4bd2ce80fe8289. It’s now merged. Thank you!